Rebis OK, Regression tests OK     Typesepecifications following %polymorphic may now contain     comment: end-of-line comment and C-style comment are both     recognized (and ignored).

+++ bisonc++/documentation/regression/fun/rulevalue/_rulevalue1.cc
@@ -19,7 +19,7 @@ RuleValue::RuleValue(int i)
     d_value.i = i;
 }
 
-RuleValue::RuleValue(size_t s)
+RuleValue::RuleValue(unsigned s)
 :
     d_tag(VARIABLE)
 {
